Rename about folder name for avoiding crash with About Components
