+++ src/components/official/details/DetailOne.js
