+++ src/components/official/details/DetailTwo.js
