+++ src/components/official/details/index.js
