+++ src/components/official/tribes/index.js
@@ -1,0 +1,4 @@
+export { default as Lyquid } from './Lyquid'
+export { default as CloudX } from './CloudX'
+export { default as Staut } from './Staut'
+export { default as RedPool } from './RedPool'
